Backend/requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.30.0
pillow==9.5.0
numpy==1.24.0
opencv-python==4.11.0.86
scikit-learn==1.3.0 